driver: rename networkusbstoragedriver.py -> usbstoragedriver.py
