+++ labgrid/driver/__init__.py
@@ -20,7 +20,7 @@ from .common import Driver
 from .qemudriver import QEMUDriver
 from .modbusdriver import ModbusCoilDriver
 from .sigrokdriver import SigrokDriver
-from .networkusbstoragedriver import NetworkUSBStorageDriver, Mode
+from .usbstoragedriver import NetworkUSBStorageDriver, Mode
 from .resetdriver import DigitalOutputResetDriver
 from .gpiodriver import GpioDigitalOutputDriver
 from .filedigitaloutput import FileDigitalOutputDriver
