+++ labgrid/driver/usbstoragedriver.py
